pyproject.toml
[project]
name = "okonomi"
version = "0.1.0"
description = "AI-powered image generation agent with creative strategies"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Dhruv Nair", email = "dhruv@huggingface.co" }]
keywords = ["ai", "image-generation", "creative-ai", "flux", "mcp"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "chromadb>=1.0.15",
  "fal-client>=0.7.0",
  "gradio>=5.38.2",
  "httpx>=0.28.1",
  "huggingface-hub>=0.33.5",
  "litellm>=1.74.8",
  "mcp[cli]>=1.12.2",
  "mcpadapt>=0.1.13",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
  "rich>=14.0.0",
  "smolagents==1.21.1",
]

[project.scripts]
okonomi = "okonomi.cli:main"

[dependency-groups]
debug = ["ipdb>=0.13.13"]
dev = ["ipdb>=0.13.13", "pyright>=1.1.403", "ruff>=0.12.5"]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
# Never enforce `E501` (line length violations).
ignore = ["C901", "E501", "E721", "E741", "F402", "F823"]
select = ["C", "E", "F", "I", "W"]

# Ignore import violations in all `__init__.py` files.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["okonomi"]

[tool.hatch.build]
include = [
  "okonomi/**/*.py",
  "okonomi/**/*.yml",
  "config.json",
]

[tool.ruff.lint.isort]
lines-after-imports = 2
known-first-party = ["okonomi"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
